+++ R/filter.R
@@ -1,38 +1,47 @@
 .filter_const_MESSAGE_FILTERED_ALL <- "All data are removed by filter."
-.filter_const_MESSAGE_LOGGER_TYP_AGG <- "Logger types can be used only for raw format."
+.filter_const_MESSAGE_LOGGER_TYPE_AGG <- "Logger types can be used only for raw format."
+.filter_const_MESSAGE_LOGGERS_AGG <- "Loggers can be used only for raw format."
 .filter_const_MESSAGE_REVERSE_COMBINATION <- "The reverse parameter can only be used for a single filter condition. Is not allowed to combine more filter conditions with reverse parameter."
+.filter_const_MESSAGE_LOGGERS_BOTH <- "It is not allowed to combine loggers and logger_types parameter."
 
 #' Filter data from myClim object
 #' @description
 #' This function filter data by localities, logger types and sensors.
 #'
 #' @details
-#' In default settings it returns the object containing input localities / logger types / sensors.
+#' In default settings it returns the object containing input localities / loggers / logger types / sensors.
 #' When you provide vector of localities e.g. `localities=c("A6W79", "A2E32")`
 #' selected localities are filtered with all loggers / sensors on those localities.
+#' When you provide vector of loggers e.g. `loggers=c("TMS_1", "TMS_2")`
+#' selected loggers are filtered through all localities.
 #' The same as When you provide vector of logger_types `logger_types=c("TMS", "TMS_L45")`
-#' selected loggers by type are filtered through all localities 
-#' (logger_types criterion is applicable only for raw data format see [myClim-package]) and
+#' selected loggers by type are filtered through all localities.
+#' (loggers or logger_types criterion is applicable only for raw data format see [myClim-package]) and
 #' the sensors parameter `sensors=c("TMS_T1", "TMS_T2")`,
 #' selected sensors are filtered through all localities.
 #' When you combine localities, logger_types and sensors, then filtering return
 #' selected sensors in selected loggers on selected localities.
 #'
-#' Parameter `reverse = TRUE` returns myClim object without listed localities, or logger types or sensors.
-#' Using  `reverse = TRUE` is not allowed for combination of localities and logger types and sensors.
-#' It is allowed to use `reverse` only with single filter criterion either locality, logger type or sensor.
+#' Parameter `reverse = TRUE` returns myClim object without listed localities or loggers or logger types or sensors.
+#' Using  `reverse = TRUE` is not allowed for combination of localities and loggers and logger types and sensors.
+#' It is allowed to use `reverse` only with single filter criterion either locality, logger, logger type or sensor.
 #' 
+#' * `reverse = TRUE` and `localities` are selected then the listed localities are removed from myClim object.
+#' * `reverse = TRUE` and `loggers` are selected then the listed loggers are removed from all localities.
 #' * `reverse = TRUE` and `logger_types` are selected then the listed logger types are removed from all localities.
-#' * `reverse = TRUE` and `localities` are selected then the listed localities are removed from myClim object. 
 #' * `reverse = TRUE` and `sensors` are selected then listed sensors are removed from all loggers / localities.
+#'
+#' Only one of parameters loggers or logger_types can be used.
 #'
 #' @template param_myClim_object
 #' @param localities locality_ids for filtering data; if NULL then do nothing (default NULL)
-#' @param logger_types types of logger for filtering data; if NULL then do nothing (default NULL). 
-#' The logger_types parameter can by used only for raw data format see [myClim-package].
 #' @param sensors sensor_names for filtering data; if NULL then do nothing see `names(mc_data_sensors)` (default NULL)
 #' @param reverse if TRUE then input localities and/or sensors are excluded (default FALSE)
 #' @param stop_if_empty if TRUE then error for empty output (default TRUE)
+#' @param logger_types types of logger for filtering data; if NULL then do nothing (default NULL).
+#' The logger_types parameter can by used only for raw data format see [myClim-package].
+#' @param loggers logger_names for filtering data; if NULL then do nothing (default NULL).
+#' The loggers parameter can by used only for raw data format see [myClim-package].
 #'
 #' @return filtered myClim object
 #' @export
@@ -55,18 +64,33 @@
 #' ## Remove "Dendro" loggers on all localities
 #' filtered_data <- mc_filter(mc_data_example_raw, logger_types="Dendro", reverse=TRUE)
 
-mc_filter <- function(data, localities=NULL, sensors=NULL, reverse=FALSE, stop_if_empty=TRUE, logger_types=NULL) {
+mc_filter <- function(data, localities=NULL, sensors=NULL, reverse=FALSE, stop_if_empty=TRUE,
+                      logger_types=NULL, loggers=NULL) {
     is_agg_format <- .common_is_agg_format(data)
     if(is_agg_format && !is.null(logger_types)){
-        stop(.filter_const_MESSAGE_LOGGER_TYP_AGG)
+        stop(.filter_const_MESSAGE_LOGGER_TYPE_AGG)
+    }
+    if(is_agg_format && !is.null(loggers)){
+        stop(.filter_const_MESSAGE_LOGGERS_AGG)
+    }
+    if(!is.null(logger_types) && !is.null(loggers)){
+        stop(.filter_const_MESSAGE_LOGGERS_AGG)
     }
 
-    if(reverse && sum(c(!is.null(localities), !is.null(sensors), !is.null(logger_types))) > 1) {
+    if(reverse && sum(c(!is.null(localities), !is.null(sensors), !is.null(logger_types), !is.null(loggers))) > 1) {
         stop(.filter_const_MESSAGE_REVERSE_COMBINATION)
     }
 
     sensors_item_function <- function(item) {
-        is_in <- is.null(logger_types) || item$metadata@type %in% logger_types
+        is_in <- FALSE
+        if(is.null(logger_types) && is.null(loggers)) {
+            is_in <- TRUE
+        } else if(!is.null(loggers)) {
+            is_in <- item$metadata@name %in% loggers
+        } else if(!is.null(logger_types)) {
+            is_in <- item$metadata@type %in% logger_types
+        }
+
         if(is.null(sensors)) {
             return(if(is_in == reverse) NULL else item)
         }
@@ -84,7 +108,7 @@ mc_filter <- function(data, localities=NULL, sensors=NULL, reverse=FALSE, stop_i
     locality_function <- function(locality) {
         is_in <- is.null(localities) || locality$metadata@locality_id %in% localities
 
-        if(is.null(sensors) && is.null(logger_types)) {
+        if(is.null(sensors) && is.null(logger_types) && is.null(loggers)) {
             return(if(is_in == reverse) NULL else locality)
         }
 
@@ -104,10 +128,10 @@ mc_filter <- function(data, localities=NULL, sensors=NULL, reverse=FALSE, stop_i
         return(locality)
     }
 
-    if(is.null(localities) && is.null(sensors) && is.null(logger_types) && !reverse) {
+    if(is.null(localities) && is.null(sensors) && is.null(logger_types) && is.null(loggers) && !reverse) {
         return(data)
     }
-    if(is.null(localities) && is.null(sensors) && is.null(logger_types) && reverse) {
+    if(is.null(localities) && is.null(sensors) && is.null(logger_types) && is.null(loggers) && reverse) {
         data$localities <- list()
     } else {
         out_localities <- purrr::map(data$localities, locality_function)